requirements.txt
# Core machine learning libraries
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Data handling and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model persistence
joblib>=1.3.0

# Progress bars (optional but helpful)
tqdm>=4.65.0
